upd js to ts
