+++ npdemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["*/**"],
+  "ext": ".ts,.json,.js",
+  "ignore": ["node_modules"],
+  "exec": "npx ts-node ./src/index.ts"
+}
