+++ src/index.ts
@@ -1,0 +1,8 @@
+import { httpServer } from "./http_server/index";
+import "dotenv/config"
+
+const HTTP_PORT = Number(process.env.HTTP_PORT) || 8181;
+const WS_PORT = Number(process.env.WS_PORT) || 3000;
+
+console.log(`Start static http server on the ${HTTP_PORT} port!`);
+httpServer.listen(HTTP_PORT);
